+++ frontend/common/ui/src/commonMain/kotlin/net/matsudamper/money/frontend/common/ui/screen/admin/user/UserSearchScreen.kt
@@ -47,7 +47,7 @@ import org.jetbrains.compose.resources.painterResource
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
-internal fun UserSearchScreen(
+public fun UserSearchScreen(
     modifier: Modifier = Modifier,
     uiState: AdminUserSearchUiState,
     snackbarHostState: SnackbarHostState,
